making exported facebook() keyword all lowercase
Perl variables, as well as Dancer-exported plugins, are most
commonly written in 'snake case', that is, all lowercase and
with underscores ("_") separating words. In case of "Facebook",
it's just a single word, so the FaceBook() exported keyword should
be renamed to 'facebook()' instead. This is also compliant to the
Dancer::Plugin::Auth::Twitter interface, which exports 'twitter()'
and not "Twitter()".