access_token => fb_access_token to play nice with others
In a large web app, chances are 'access_token' is already used
as session key by either the app itself or some ill-behaved plugin.
So, in order to play nice with other plugins and avoid namespace
clash, we changed the (internal) session('access_token') key to
session('fb_access_token').